user/requirements.txt
# FastAPI and related packages
fastapi==0.110.0
uvicorn==0.28.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.2
Faker==25.8.0
uuid==1.30
psycopg2==2.9.9
python-jose==3.3.0
PyJWT==2.8.0

# Testing packages
pytest==7.2.0
pytest-asyncio==0.20.0
httpx==0.23.0